+++ docs/changelog.md
@@ -7,6 +7,8 @@
 
 ### Bugfixes
 - Workaround for MRO resolution for remote module connections
+- Fixed client crashing when server disconnects for remote module connections by introduncing a new state for module- DISCONNECTED
+- Disabled deactivation of remote hardware module from client for remote module connections
 
 ### New Features
 - Added an new `ConnectorList` type for connecting an unspecified number of modules of the same interface to a module.